pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "potato-bot"
version = "3.1.0"
description = "Discord 多功能管理機器人"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Potato Bot Team", email = "support@potato-bot.com"}
]

dependencies = [
    # Core Discord Bot
    "discord.py>=2.3.2",

    # Database & Caching
    "aiomysql>=0.2.0",

    # Data Serialization
    "orjson>=3.9.9",
    "msgpack>=1.0.7",

    # Environment & Configuration
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",

    # Security & Encryption
    "cryptography>=41.0.7",
    "PyNaCl>=1.5.0",

    # HTTP Client
    "aiohttp>=3.9.0",

    # File & Media Processing
    "aiofiles>=24.1.0",
    "Pillow>=10.0.0",
    "yt-dlp>=2024.3.10",

    # Utilities
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "regex>=2023.10.3",
    "psutil>=5.9.6",
]

[tool.setuptools]
packages = ["bot", "shared", "src"]
package-dir = {"" = "."}
include-package-data = true
